+++ ci/requirements/install_ci_envs.sh
@@ -27,9 +27,9 @@ create_this_env () {
 	/bin/rm temp_env.yml > /dev/null 2>&1
 }
 
-create_this_env "argopy-tests-py37dev" "py3.7-dev.yml"
+#create_this_env "argopy-tests-py37dev" "py3.7-dev.yml"
 #create_this_env "argopy-tests-py38free-small" "py3.8-small-free.yml"
-#create_this_env "argopy-tests-py38free" "py3.8-free.yml"
+create_this_env "argopy-tests-py38free" "py3.8-free.yml"
 #create_this_env "argopy-tests-py38dev" "py3.8-dev.yml"
 
 exit 0